pyproject.toml
[project]
name = "anonshield"
version = "0.1.0"
description = "Anonimizador"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "email-validator>=2.2.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "presidio-analyzer>=2.2.357",
    "presidio-anonymizer>=2.2.357",
    "protobuf>=6.30.1",
    "python-docx>=1.1.2",
    "sentencepiece>=0.2.0",
    "transformers>=5.3.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.4.0",
    "opencv-python-headless>=4.10.0",
    "PyMuPDF>=1.24.9",
    "lxml>=5.2.2",
    "ijson>=3.4.0.post0",
    "orjson>=3.11.4",
    "tenacity",
    "PyYAML>=6.0.1",
    "python-magic>=0.4.27",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    "tabulate>=0.9.0",
    "grex>=1.0.1",
    "wordcloud>=1.9.3",
    "sentence-transformers>=3.0.0",
    "hdbscan>=0.8.33",
    "scipy>=1.13.0",
    "statsmodels>=0.14.0",
    "en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl",
    "torch>=2.6.0",
    "torchvision",
    "cupy-cuda12x==12.3.0",
]

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[project.optional-dependencies]
easyocr = ["easyocr>=1.7.0"]
paddleocr = ["paddleocr>=2.8.0", "paddlepaddle>=2.6.0"]
doctr = ["python-doctr>=1.0.0"]
onnxtr = ["onnxtr[gpu]>=0.6.0"]
onnxtr-cpu = ["onnxtr[cpu]>=0.6.0"]
kerasocr = ["keras-ocr>=0.9.3"]
surya = ["surya-ocr>=0.6.0"]
rapidocr = ["rapidocr-onnxruntime>=1.3.0"]
glm-ocr = ["torch>=2.0.0"]
lighton-ocr = ["torch>=2.0.0", "pypdfium2>=4.0.0"]
paddle-vl = ["torch>=2.0.0"]
deepseek-ocr = ["einops>=0.8.0", "addict>=2.4.0", "easydict>=1.13"]
monkey-ocr = ["huggingface_hub>=0.24.0"]
chandra-ocr = ["transformers>=5.3.0"]
dots-ocr = ["transformers>=5.3.0"]
qwen-vl = ["transformers>=5.3.0", "qwen-vl-utils>=0.0.8", "accelerate>=0.30.0"]
ocr-all = [
    "easyocr>=1.7.0",
    "python-doctr>=1.0.0",
    "onnxtr[gpu]>=0.6.0",
    "surya-ocr>=0.6.0",
    "rapidocr-onnxruntime>=1.3.0",
    "torch>=2.0.0",
    "pypdfium2>=4.0.0",
    "einops>=0.8.0",
    "huggingface_hub>=0.24.0",
    "qwen-vl-utils>=0.0.8",
    "accelerate>=0.30.0",
]
paddle-ocr = [
    "paddleocr>=3.4.0",
    "paddlepaddle>=3.3.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-timeout>=2.4.0",
]
# Web server + infrastructure (FastAPI, Celery, Redis…)
# Used by Dockerfile.dev and Dockerfile (prod)
web = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "celery[redis]>=5.4.0",
    "redis>=5.2.0",
    "python-multipart>=0.0.12",
    "httpx>=0.27.0",
    "aiofiles>=24.1.0",
    "slowapi>=0.1.9",
    "pt-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_lg-3.8.0/pt_core_news_lg-3.8.0-py3-none-any.whl",
]
# Web integration tests (httpx test client)
web-dev = [
    { include-group = "web" },
    "pytest>=9.0.3",
    "pytest-timeout>=2.4.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m not slow')",
    "slm: marks tests that require Ollama running (deselect with '-m not slm')",
    "ocr: marks tests that require OCR engines to be installed",
]
